refactor: update kdfType filename
