+++ apps/web/src/app/settings/change-kdf/change-kdf.component.ts
@@ -9,7 +9,7 @@ import {
   DEFAULT_ARGON2_MEMORY,
   DEFAULT_ARGON2_PARALLELISM,
   KdfType,
-} from "@bitwarden/common/enums/kdfType";
+} from "@bitwarden/common/enums/kdf-type";
 import { DialogService } from "@bitwarden/components";
 
 import { ChangeKdfConfirmationComponent } from "./change-kdf-confirmation.component";
