+++ libs/common/src/auth/models/request/set-key-connector-key.request.ts
@@ -1,4 +1,4 @@
-import { KdfType } from "../../../enums/kdfType";
+import { KdfType } from "../../../enums/kdf-type";
 import { KeysRequest } from "../../../models/request/keys.request";
 import { KdfConfig } from "../domain/kdf-config";
