+++ libs/common/src/models/request/register.request.ts
@@ -1,5 +1,5 @@
 import { CaptchaProtectedRequest } from "../../auth/models/request/captcha-protected.request";
-import { KdfType } from "../../enums/kdfType";
+import { KdfType } from "../../enums/kdf-type";
 
 import { KeysRequest } from "./keys.request";
 import { ReferenceEventRequest } from "./reference-event.request";
